Work around kyua bug
If a cmake test is added for each atf test, with kyua being invoked once
for each test, then running ctest with any level of parallelism (-j8 for
example) introduces a chance that the test run fails because kyua tries
to write to a db that is locked by another test.

To work around this, this commit folds all of the atf tests into a
single cmake test so that kyua is invoked only once.